MLB Client Software
Developed as a class project for SJSU CMPE 138 (Database Design and Architecture). This is a Python Tkinter based GUI designed to be used with a MSSQL database (defined in the SQL file). Project was intended to provide a UI for adding/updating/querying data about Major League Baseball such as players, teams, stadiums, leagues, etc.