Spartan Smart Board Project
Developed as a senior project for SJSU CMPE 195 (Senior Design Project). This is a Python based project that utilizes the wxPython, openCV, and piCamera libraries to interface with a Raspberry Pi 2 development board and a Raspberry Pi NoIR camera to create an interactive "whiteboard" to be projected onto a wall and interfaced with by users holding an IR LED pen. Project was intended to be run on a Raspberry Pi with a PiNoIR camera and a projector/monitor connected in order to allow users to interact with a desktop based on IR LED pen movements. The tracking was done using openCV to process IR LEDs being captured in images taken by the IR filter-free PiNoIR camera. The images are then processed into mouse movements so that a user could have an interactive whiteboard if they used the system's paint program or could be used to interact with other software on the screen such as to play a game of connect 4 or solitaire.